tests: move pageserver http client to separate module
